+++ CHANGELOG.md
@@ -7,9 +7,9 @@ v0.6.0 (unreleased)
 -   **Breaking**: Filter & Interceptor interfaces are renamed
     UnaryInboundMiddleware & UnaryOutboundMiddleware respectively.
 -   **Breaking**: yarpc.Config doesn't take Filter & Interceptor anymore but
-    InboundMiddlewares and OutboundMiddlewares.
--   Oneway middlewares are now available via OnewayInboundMiddleware
-    OnewayOutboundMiddlewares interfaces.
+    InboundMiddleware and OutboundMiddleware.
+-   Oneway middleware are now available via OnewayInboundMiddleware
+    OnewayOutboundMiddleware interfaces.
 
 
 v0.5.0 (2016-11-21)